bot_scrap_svpressa/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
fake-useragent==1.4.0
frozenlist==1.4.1
idna==3.6
lxml==5.1.0
magic-filter==1.0.12
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
soupsieve==2.5
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
